fix: resolve build failures and lint errors in NPC creation workflow
- Remove unused variables and imports in NPCCreationForm
- Fix prop names in tab components (onUpdate vs onUpdateFormData)
- Add missing suggestedAC prop to NPCStatsTab
- Remove errors prop from NPCDetailsTab and add missing event handlers
- Fix trailing whitespace in test-helpers.ts
- Prefix unused function parameters with underscore
